cleaned up networking
